+++ string_builder.c
@@ -31,7 +31,7 @@ char *string_builder_get(StringBuilder *this)
     return ret;
 }
 
-size_t string_builder_size(StringBuilder *this)
+int string_builder_size(StringBuilder *this)
 {
     return vector_size(this->data) + 1;
 }
